requirements.txt
asgiref==3.6.0
Django==3.2.16
django-cors-headers==3.13.0
django-extra-fields==3.0.2
djangorestframework==3.14.0
gunicorn==20.1.0
numpy==1.21.6
opencv-python==4.6.0.66
Pillow==9.2.0
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.4.0
Unipath==1.1
whitenoise==6.3.0
